backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.7.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
faiss-cpu==1.13.0
